# Short-circuit node comparisons, add more heuristics