A bit slower with more heuristics, but better. Seems worth it.

**Before**

```
 Benchmarking: anytime_beam_search(board, {'power': 10})
 ....................................................................................................
 Solutions of avg. length 15.130 in 12.871 sec.
 Avg. length times sec (lower is better) => 194.738

Benchmarking: heuristic_search(board, {'max_nodes': 128000})
 ....................................................................................................
 Solutions of avg. length 18.160 in 8.770 sec.
 Avg. length times sec (lower is better) => 159.261

Benchmarking: monte_carlo_search(board, {'iterations': 3450, 'seed': 1})
 ....................................................................................................
 Solutions of avg. length 16.640 in 16.079 sec.
 Avg. length times sec (lower is better) => 267.548
```

**After**

```
 Benchmarking: anytime_beam_search(board, {'power': 10})
 ....................................................................................................
 Solutions of avg. length 15.110 in 15.676 sec.
 Avg. length times sec (lower is better) => 236.858

Benchmarking: heuristic_search(board, {'max_nodes': 128000})
 ....................................................................................................
 Solutions of avg. length 17.820 in 14.402 sec.
 Avg. length times sec (lower is better) => 256.647

Benchmarking: monte_carlo_search(board, {'iterations': 3450, 'seed': 1})
 ....................................................................................................
 Solutions of avg. length 16.390 in 16.623 sec.
 Avg. length times sec (lower is better) => 272.457
```